build: envsetup: propogate return code in mka
Return codes were not properly returned back to the shell when using
mka, so it was impossible to determine when the make command failed to
execute. Reading the return code is critical in automated build
environments, where the status of the build is dependent on the return
code. This commit returns the return code to the caller so that it can
be read as needed.

Change-Id: I67e962382abf2d1423ca280228c6c4c3cd3e4f86
Signed-off-by: Evan Anderson <evananderson@aospa.co>
(cherry picked from commit 39511a50cb6915b2b7b9ca87df3da6d4518e126c)